Simplify eserv.c
eserv.c was calling recv_packet() only once, and was the only caller.
Thus, recv_packet() was moved from eutils.c to eserv.c.
Since recv_packet() was very simple, it was also merged with echo().

Fixed typecast in the call of bind() as well as its sizeof parameter.

Call of recvfrom() in main() was corrected.