Utilize callback functions for processing files
Move definition of struct fc_info to eutils.h and include in the struct a callback function that deals with processing files.

By including callback functions, we can encapsulate the corking code in eclicork.c. eutils.c does not have to know about the number of currently corked bytes or any of the (un)corking operations.

In essence, this keeps the code efficient by utilizing functions in eutils for common operations, and only referring back to corking operations when absolutely necessary. Including the callback function syntax in the struct fc_info also keeps the function signatures simple.